server/requirements.txt
Flask
python-dotenv
huggingface-hub
Pillow
pretty_midi
mido
note-seq>=0.0.5